Change method input list
